Remove unused ValidateAuthToken flow and messages
