+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/apps/users/GetUsersMeetingReqMsgHdlr.scala
@@ -1,7 +1,9 @@
 package org.bigbluebutton.core.apps.users
 
-import org.bigbluebutton.common2.msgs.GetUsersMeetingReqMsg
+import org.bigbluebutton.common2.msgs.{ GetUsersMeetingReqMsg, VoiceConfUser, WebUser }
+import org.bigbluebutton.core.models.{ Users2x, VoiceUsers }
 import org.bigbluebutton.core.running.{ HandlerHelpers, LiveMeeting, OutMsgRouter }
+import org.bigbluebutton.core2.message.senders.MsgBuilder
 
 trait GetUsersMeetingReqMsgHdlr extends HandlerHelpers {
   this: UsersApp =>
@@ -13,4 +15,28 @@ trait GetUsersMeetingReqMsgHdlr extends HandlerHelpers {
     sendAllUsersInMeeting(msg.body.userId)
     sendAllVoiceUsersInMeeting(msg.body.userId, liveMeeting.voiceUsers, liveMeeting.props.meetingProp.intId)
   }
+
+  def sendAllUsersInMeeting(requesterId: String): Unit = {
+    val meetingId = liveMeeting.props.meetingProp.intId
+    val users = Users2x.findAll(liveMeeting.users2x)
+    val webUsers = users.map { u =>
+      WebUser(intId = u.intId, extId = u.extId, name = u.name, role = u.role,
+        guest = u.guest, authed = u.authed, guestStatus = u.guestStatus,
+        locked = u.locked, presenter = u.presenter, avatar = u.avatar, clientType = u.clientType)
+    }
+
+    val event = MsgBuilder.buildGetUsersMeetingRespMsg(meetingId, requesterId, webUsers)
+    outGW.send(event)
+  }
+
+  def sendAllVoiceUsersInMeeting(requesterId: String, voiceUsers: VoiceUsers, meetingId: String): Unit = {
+    val vu = VoiceUsers.findAll(voiceUsers).map { u =>
+      VoiceConfUser(intId = u.intId, voiceUserId = u.voiceUserId, callingWith = u.callingWith, callerName = u.callerName,
+        callerNum = u.callerNum, color = u.color, muted = u.muted, talking = u.talking, listenOnly = u.listenOnly)
+    }
+
+    val event = MsgBuilder.buildGetVoiceUsersMeetingRespMsg(meetingId, requesterId, vu)
+    outGW.send(event)
+  }
+
 }
